Move AggregatedMetricsSet to metrics for further reuse
